+++ lib/kafka/protocol/fetch_request.rb
@@ -30,6 +30,10 @@ module Kafka
         FETCH_API
       end
 
+      def api_version
+        2
+      end
+
       def response_class
         Protocol::FetchResponse
       end